# packages/homeflow/openapi.yaml
openapi: 3.1.0
info:
  title: HomeFlow — Extropy Engine IoT Smart Home Service
  version: 1.0.0
  description: |
    HomeFlow is an IoT-integrated smart home/building management service
    built on top of the Extropy Engine. It measures thermodynamic entropy
    reduction from home automation actions and mints XP through the
    verification loop.

    Core formula: XP = R × F × ΔS × (w · E) × log(1/Tₛ)

    HomeFlow operates in the THERMODYNAMIC entropy domain and interoperates
    with all 13+ ecosystem apps via the shared EventPayloadMap and a
    plugin/adapter interoperability layer.
  contact:
    name: Extropy Engine
    url: https://github.com/00ranman/extropy-engine
  license:
    name: MIT
servers:
  - url: http://localhost:4015
    description: Local development
  - url: http://homeflow:4015
    description: Docker Compose internal
...
paths:
...
  /events:
    post:
      tags: [Events]
      summary: Receive inter-service events (webhook)
      description: |
        Standardized webhook endpoint for receiving events from any
        ecosystem service. Supports all EventPayloadMap event types
        plus HomeFlow-specific events.
      operationId: receiveEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomingWebhookEvent'
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  processed:
                    type: boolean
                  crossDomainData:
                    $ref: '#/components/schemas/CrossDomainEntropyData'
                  aggregation:
                    type: object
                    properties:
                      compositeDeltaS:
                        type: number
                      compositeXP:
                        type: number
        '400':
          description: Invalid event format
...
  /api/v1/governance/dfao:
    post:
      tags: [Governance]
      summary: Create a household DFAO
      description: |
        Creates a NANO-scale DFAO for the household via the DFAO Registry.
        The DFAO governs energy policies, automation rules, and contribution
        distribution.
      operationId: createHouseholdDFAO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [householdId, householdName, founderValidatorId]
              properties:
                householdId:
                  type: string
                householdName:
                  type: string
                founderValidatorId:
                  type: string
      responses:
        '201':
          description: DFAO created
  /api/v1/schedules:
    post:
      tags: [Schedules]
      summary: Create an automation schedule
      operationId: createSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Schedule created
    get:
      tags: [Schedules]
      summary: List schedules
      operationId: listSchedules
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedules list
  /api/v1/tokens/{householdId}:
    get:
      tags: [Tokens]
      summary: Get token balances for household
      operationId: getTokenBalances
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token balances
  /api/v1/credentials/{householdId}:
    get:
      tags: [Credentials]
      summary: Get credentials for household
      operationId: getCredentials
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credentials list
  /api/v1/reputation/{validatorId}:
    get:
      tags: [Reputation]
      summary: Get THERMODYNAMIC reputation for validator
      operationId: getReputation
      parameters:
        - name: validatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reputation data
  /api/v1/dag/{householdId}:
    get:
      tags: [DAG]
      summary: Get DAG vertex references for household
      operationId: getDAGReferences
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DAG references
...
  /api/v1/interop/adapters:
    get:
      tags: [Interop]
      summary: List registered ecosystem adapters
      operationId: listAdapters
      responses:
        '200':
          description: Adapters list
    post:
      tags: [Interop]
      summary: Register a new ecosystem adapter at runtime
      operationId: registerAdapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appId, appName]
              properties:
                appId:
                  type: string
                appName:
                  type: string
                entropyDomains:
                  type: array
                  items:
                    type: string
                publishedEvents:
                  type: array
                  items:
                    type: string
                subscribedEvents:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Adapter registered
  /api/v1/interop/aggregations:
    get:
      tags: [Interop]
      summary: Get cross-domain aggregation history
      operationId: getAggregations
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Aggregation history
  /api/v1/temporal/patterns:
    get:
      tags: [Schedules]
      summary: Detect seasonal patterns
      operationId: detectPatterns
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detected patterns
components:
  schemas:
...
    IncomingWebhookEvent:
      type: object
      required: [eventId, type, source, payload, correlationId, timestamp, version]
      properties:
        eventId:
          type: string
        type:
          type: string
        source:
          type: string
        payload:
          type: object
        correlationId:
          type: string
        timestamp:
          type: string
        version:
          type: integer
    CrossDomainEntropyData:
      type: object
      nullable: true
      properties:
        sourceApp:
          type: string
        sourceDomain:
          type: string
        sourceDeltaS:
          type: number
        timestamp:
          type: string
...
    CreateScheduleRequest:
      type: object
      required: [householdId, name, type, actions]
      properties:
        householdId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [time_based, occupancy, seasonal, solar_aware, price_aware]
        cronExpression:
          type: string
        seasonId:
          type: string
        conditions:
          type: array
          items:
            type: object
        actions:
          type: array
          items:
            type: object
            properties:
              deviceId:
                type: string
              commandType:
                type: string
              parameters:
                type: object
...